Create Censoring.R
Testing independece of X and Y under left truncation and right censoring using only uncensored data.
The resulting weight function is I(x<y)S(y-x) where S is the survival function of censoring.
The simulation compares the use of the true S to Kaplan-Meier and Neleson-A'alen estimates of S.